requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
PyJWT==2.6.0
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
pytz==2022.7.1
openpyxl==3.1.2
Pillow==9.5.0
django-debug-toolbar==4.1.0
django-redis==5.3.0
django-redis-cache==3.0.1
redis==4.6.0
